modify user, compile ok

+++ miniogw/gateway.go
@@ -29,7 +29,8 @@ var (
 	errDeleteObjects        = errors.New("error(s) occurred while deleting objects")
 )
 
-func Start(addr, pwd string) error {
+// Start gateway
+func Start(addr, pwd, endPoint string) error {
 	minio.RegisterGatewayCommand(cli.Command{
 		Name:            "lfs",
 		Usage:           "Mefs Log File System Service (LFS)",
@@ -53,7 +54,7 @@ func Start(addr, pwd string) error {
 	// "gateway" represents gatewat mode; respective, "server" represents server mode
 	// "lfs" is subcommand, should equal to RegisterGatewayCommand{Name}
 	go minio.Main([]string{"memoriae", "gateway", "lfs",
-		"--address", "127.0.0.1:5080", "--config-dir", gwConf})
+		"--address", endPoint, "--config-dir", gwConf})
 
 	return nil
 }
@@ -63,7 +64,7 @@ func mefsGatewayMain(ctx *cli.Context) {
 	minio.StartGateway(ctx, &Mefs{"lfs"})
 }
 
-// LFS implements Gateway.
+// Mefs implements Lfs Gateway.
 type Mefs struct {
 	host string
 }
@@ -109,21 +110,21 @@ func (l *lfsGateway) StorageInfo(ctx context.Context) (si minio.StorageInfo) {
 	si.Backend.Type = minio.BackendGateway
 	si.Backend.GatewayOnline = true
 
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
 		return si
 	}
 
-	si.Used, _ = lfs.ShowStorageSpaceAll()
-
+	use, _ := lfs.ShowStorage()
+	si.Used = []uint64{use}
 	return si
 }
 
 // MakeBucketWithLocation creates a new container on LFS backend.
 func (l *lfsGateway) MakeBucketWithLocation(ctx context.Context, bucket, options string) error {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return user.ErrLfsServiceNotReady
 	}
 	bucketOptions := &pb.BucketOptions{}
 	err := json.Unmarshal([]byte(options), bucketOptions)
@@ -137,9 +138,9 @@ func (l *lfsGateway) MakeBucketWithLocation(ctx context.Context, bucket, options
 
 // GetBucketInfo gets bucket metadata.
 func (l *lfsGateway) GetBucketInfo(ctx context.Context, bucket string) (bi minio.BucketInfo, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return bi, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return bi, user.ErrLfsServiceNotReady
 	}
 	bucketInfo, err := lfs.HeadBucket(bucket)
 	if err != nil {
@@ -152,11 +153,11 @@ func (l *lfsGateway) GetBucketInfo(ctx context.Context, bucket string) (bi minio
 
 // ListBuckets lists all LFS buckets.
 func (l *lfsGateway) ListBuckets(ctx context.Context) (buckets []minio.BucketInfo, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return nil, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return nil, user.ErrLfsServiceNotReady
 	}
-	bucketsInfo, err := lfs.ListBucket("")
+	bucketsInfo, err := lfs.ListBuckets("")
 	if err != nil {
 		return nil, err
 	}
@@ -170,9 +171,9 @@ func (l *lfsGateway) ListBuckets(ctx context.Context) (buckets []minio.BucketInf
 
 // DeleteBucket deletes a bucket on LFS.
 func (l *lfsGateway) DeleteBucket(ctx context.Context, bucket string) error {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return errLfsServiceNotReady
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return user.ErrLfsServiceNotReady
 	}
 
 	_, err := lfs.DeleteBucket(bucket)
@@ -182,13 +183,16 @@ func (l *lfsGateway) DeleteBucket(ctx context.Context, bucket string) error {
 
 // ListObjects lists all blobs in LFS bucket filtered by prefix.
 func (l *lfsGateway) ListObjects(ctx context.Context, bucket, prefix, marker, delimiter string, maxKeys int) (loi minio.ListObjectsInfo, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return loi, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return loi, user.ErrLfsServiceNotReady
 	}
-	objs, _, err := lfs.ListObject(bucket, prefix, false)
+
+	var ops user.ObjectOptions
+
+	objs, err := lfs.ListObjects(bucket, prefix, ops)
 	if err != nil {
-		return loi, user.ErrLfsIsNotRunning
+		return loi, user.ErrLfsServiceNotReady
 	}
 	loi.Objects = make([]minio.ObjectInfo, len(objs))
 	for i, v := range objs {
@@ -204,13 +208,16 @@ func (l *lfsGateway) ListObjects(ctx context.Context, bucket, prefix, marker, de
 // ListObjectsV2 lists all blobs in LFS bucket filtered by prefix
 func (l *lfsGateway) ListObjectsV2(ctx context.Context, bucket, prefix, continuationToken, delimiter string, maxKeys int,
 	fetchOwner bool, startAfter string) (loi minio.ListObjectsV2Info, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return loi, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return loi, user.ErrLfsServiceNotReady
 	}
-	objs, _, err := lfs.ListObject(bucket, prefix, false)
+
+	var ops user.ObjectOptions
+
+	objs, err := lfs.ListObjects(bucket, prefix, ops)
 	if err != nil {
-		return loi, user.ErrLfsIsNotRunning
+		return loi, user.ErrLfsServiceNotReady
 	}
 	loi.Objects = make([]minio.ObjectInfo, len(objs))
 	for i, v := range objs {
@@ -225,18 +232,16 @@ func (l *lfsGateway) ListObjectsV2(ctx context.Context, bucket, prefix, continua
 
 // GetObjectNInfo - returns object info and locked object ReadCloser
 func (l *lfsGateway) GetObjectNInfo(ctx context.Context, bucket, object string, rs *minio.HTTPRangeSpec, h http.Header, lockType minio.LockType, opts minio.ObjectOptions) (gr *minio.GetObjectReader, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return nil, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return gr, user.ErrLfsServiceNotReady
 	}
+
 	objInfo, err := l.GetObjectInfo(ctx, bucket, object, opts)
 	if err != nil {
 		return nil, err
 	}
-	start, length, err := rs.GetOffsetLength(objInfo.Size)
-	if err != nil {
-		return nil, err
-	}
+
 	piper, pipew := io.Pipe()
 	bufw := bufio.NewWriterSize(pipew, user.DefaultBufSize)
 	checkErrAndClosePipe := func(err error) error {
@@ -249,14 +254,10 @@ func (l *lfsGateway) GetObjectNInfo(ctx context.Context, bucket, object string, 
 	}
 	var complete []user.CompleteFunc
 	complete = append(complete, checkErrAndClosePipe)
-	dl, err := lfs.ConstructDownload(bucket, object, bufw, complete, &user.DownloadOptions{
-		Start:  start,
-		Length: length,
-	})
+	err = lfs.GetObject(bucket, object, bufw, complete, user.DefaultDownloadOptions())
 	if err != nil {
 		return gr, err
 	}
-	go dl.Start(ctx)
 	// Setup cleanup function to cause the above go-routine to
 	// exit in case of partial read
 	pipeCloser := func() { piper.Close() }
@@ -270,9 +271,9 @@ func (l *lfsGateway) GetObjectNInfo(ctx context.Context, bucket, object string, 
 // startOffset indicates the starting read location of the object.
 // length indicates the total length of the object.
 func (l *lfsGateway) GetObject(ctx context.Context, bucket, key string, startOffset, length int64, writer io.Writer, etag string, opts minio.ObjectOptions) error {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return user.ErrLfsServiceNotReady
 	}
 
 	var errRes error
@@ -283,25 +284,25 @@ func (l *lfsGateway) GetObject(ctx context.Context, bucket, key string, startOff
 	}
 	var complete []user.CompleteFunc
 	complete = append(complete, checkErrAndClosePipe)
-	dl, err := lfs.ConstructDownload(bucket, key, bufw, complete, &user.DownloadOptions{
+	err := lfs.GetObject(bucket, key, bufw, complete, &user.DownloadOptions{
 		Start:  startOffset,
 		Length: length,
 	})
 	if err != nil {
 		return err
 	}
-	dl.Start(ctx)
 	return errRes
 }
 
 // GetObjectInfo reads object info and replies back ObjectInfo.
 func (l *lfsGateway) GetObjectInfo(ctx context.Context, bucket, object string, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return minio.ObjectInfo{}, user.ErrLfsIsNotRunning
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return objInfo, user.ErrLfsServiceNotReady
 	}
 
-	obj, _, err := lfs.HeadObject(bucket, object, false)
+	var ops user.ObjectOptions
+	obj, err := lfs.HeadObject(bucket, object, ops)
 	if err != nil {
 		return minio.ObjectInfo{}, err
 	}
@@ -319,24 +320,25 @@ func (l *lfsGateway) GetObjectInfo(ctx context.Context, bucket, object string, o
 
 // PutObject creates a new object with the incoming data.
 func (l *lfsGateway) PutObject(ctx context.Context, bucket, object string, r *minio.PutObjReader, opts minio.ObjectOptions) (objInfo minio.ObjectInfo, err error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return objInfo, errLfsServiceNotReady
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return objInfo, user.ErrLfsServiceNotReady
 	}
 
 	reader := bufio.NewReaderSize(r.Reader, user.DefaultBufSize)
-	ulJob, err := lfs.ConstructUpload(object, "", bucket, reader)
+	obj, err := lfs.PutObject(bucket, object, reader)
 	if err != nil {
 		return objInfo, err
 	}
 
-	// upload
-	err = ulJob.Start(ctx)
-	if err != nil {
-		return objInfo, err
+	objInfo = minio.ObjectInfo{
+		Bucket:      bucket,
+		Name:        object,
+		IsDir:       obj.Dir,
+		ETag:        obj.ETag,
+		ContentType: obj.ContentType,
+		Size:        obj.Size,
 	}
-
-	objInfo, err = l.GetObjectInfo(ctx, bucket, object, opts)
 
 	return objInfo, err
 }
@@ -348,9 +350,9 @@ func (l *lfsGateway) CopyObject(ctx context.Context, srcBucket, srcObject, dstBu
 
 // DeleteObject deletes a blob in bucket.
 func (l *lfsGateway) DeleteObject(ctx context.Context, bucket, object string) error {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return errLfsServiceNotReady
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return user.ErrLfsServiceNotReady
 	}
 
 	_, err := lfs.DeleteObject(bucket, object)
@@ -359,9 +361,9 @@ func (l *lfsGateway) DeleteObject(ctx context.Context, bucket, object string) er
 }
 
 func (l *lfsGateway) DeleteObjects(ctx context.Context, bucket string, objects []string) ([]error, error) {
-	lfs := user.GetLfsService(l.userID)
-	if lfs == nil {
-		return nil, errLfsServiceNotReady
+	lfs := user.GetUser(l.userID)
+	if lfs == nil || !lfs.Online() {
+		return nil, user.ErrLfsServiceNotReady
 	}
 
 	errFlag := 0
